+++ package.json
@@ -17,10 +17,12 @@
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.52.1",
     "react-icons": "^5.2.1",
+    "react-loader-spinner": "^6.1.6",
     "react-minimal-pie-chart": "^8.4.0",
     "react-router-dom": "^6.24.1",
     "react-scroll": "^1.9.0",
-    "react-tilt": "^1.0.2"
+    "react-tilt": "^1.0.2",
+    "react-toastify": "^10.0.5"
   },
   "devDependencies": {
     "@types/react": "^18.3.3",
