+++ src/main.jsx
@@ -2,9 +2,21 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
+import { Bounce, ToastContainer } from 'react-toastify'
+import 'react-toastify/dist/ReactToastify.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <App />
+    <ToastContainer
+position="top-right"
+autoClose={3000}
+hideProgressBar={false}
+closeOnClick
+draggable
+pauseOnHover
+theme="light"
+transition= {Bounce}
+/>
   </React.StrictMode>,
 )
