+++ src/services/auth.js
@@ -8,6 +8,6 @@ export const apiLogin = async(payload) => {
     return apiClient.post("/auth/login", payload);
 };
 
-export const apiCheckIfUsernameExists = async(userName) => {
-    return apiCheckIfUsernameExists.get(`/auth/${userName}`);
+export const apiCheckUsernameExists = async(userName) => {
+    return apiClient.get(`/auth/${userName}`);
 };
